only initialise functions in the top level of the module
