+++ lib/index.js.flow
@@ -16,6 +16,8 @@ export type GetRouteNameFromResourceType = (resource: string) => ?RouteNameType 
 export type GetAdditionalStateType = (routeName: RouteNameType, resource: string) => StateAdditionsType;
 type GetElementType = (querySelector: string) => Promise<?Element>;
 
+export type LifecycleType = AsyncGenerator<true, boolean, boolean>;
+
 export type OptionsType = {
   importTimeout: number,
   getRouteNameFromResource: GetRouteNameFromResourceType,
