+++ lib/utils/script.js
@@ -1,10 +1,18 @@
 // @flow
 
+import AppManagerError from './app-manager-error';
+import { delay } from './config';
+import { levels } from '../constants';
+
+import type Context from './context';
+
 import type {
   StateWithAdditionsType,
   LoadedScriptType,
   ScriptType,
   StatusDetailsType,
+  FragmentsMapType,
+  FragmentNameType,
 } from '../index';
 
 function getV5State(state) {
@@ -58,7 +66,7 @@ function getV5State(state) {
   };
 }
 
-export default function wrapScript(script: LoadedScriptType): ScriptType {
+export function wrapScript(script: LoadedScriptType): ScriptType {
   return {
     async hydrate(slotElement: Element, state: StateWithAdditionsType) {
       switch (script.version) {
@@ -267,5 +275,58 @@ export default function wrapScript(script: LoadedScriptType): ScriptType {
         }
       }
     },
+  };
+}
+
+export default function initGetFragmentScript(fragments: FragmentsMapType, importTimeout: number, context: Context) {
+  const cachedScripts: Map<FragmentNameType, ?ScriptType> = new Map();
+
+  return async function getFragmentScript(fragmentName: FragmentNameType): Promise<?ScriptType> {
+    const cachedScript = cachedScripts.get(fragmentName);
+
+    if (cachedScript) {
+      return cachedScript;
+    }
+
+    const errData = {
+      source: 'load_script',
+      fragment: fragmentName,
+      slot: null,
+    };
+
+    const { loadScript } = fragments[fragmentName];
+
+    let wrappedScript = null;
+
+    if (typeof loadScript === 'function') {
+      let script;
+
+      const state = context.getState();
+
+      try {
+        [script] = await Promise.all([loadScript(state), delay(importTimeout)]);
+      } catch (err) {
+        throw new AppManagerError(err, {
+          ...errData,
+          level: levels.ERROR,
+          code: 'load_script',
+          recoverable: true,
+        });
+      }
+
+      if (!script) {
+        throw new AppManagerError(`loadScript timed out for fragment ${fragmentName}.`, {
+          ...errData,
+          level: levels.ERROR,
+          code: 'time_out',
+          recoverable: true,
+        });
+      }
+
+      wrappedScript = wrapScript(script);
+    }
+
+    cachedScripts.set(fragmentName, wrappedScript);
+    return wrappedScript;
   };
 }
